chore: release version v0.4.3.5 with benchmark screen crash hotfix and microarchitectural bottleneck optimizations

+++ app/src/main/java/com/example/relab_tool/benchmark/domain/model/PillarScore.kt
@@ -2,7 +2,7 @@ package com.example.relab_tool.benchmark.domain.model
 
 data class PillarScore(
     val pillar: BenchmarkPillar,
-    val score: Int,
+    val score: Double,
     val subScores: List<SubScore>,
     val isSkipped: Boolean = false
 )
